+++ src/library/policy/upgrades/upgrades.py
@@ -67,3 +67,8 @@ def update_faceted_layout(context):
         obj = brain.getObject()
         layout = FacetedLayout(obj)
         layout.update_layout(layout="faceted-map")
+
+
+def set_banner_scale(context=None):
+    # Set the default scale for the banner
+    api.portal.set_registry_record("collective.behavior.banner.browser.controlpanel.IBannerSettingsSchema.banner_scale", "banner")